[Changed] Make body required in .toHaveBeenFetchedWith() matcher
